+++ src/wf_psf/psf_models/models/psf_model_semiparametric.py
@@ -10,9 +10,9 @@ to manage the parameters of the psf semi-parametric model.
 import numpy as np
 import tensorflow as tf
 from wf_psf.psf_models import psf_models as psfm
-from wf_psf.psf_models import tf_layers as tfl
+from wf_psf.psf_models.tf_modules import tf_layers as tfl
 from wf_psf.utils.utils import decompose_tf_obscured_opd_basis
-from wf_psf.psf_models.tf_layers import (
+from wf_psf.psf_models.tf_modules.tf_layers import (
     TFBatchPolychromaticPSF,
     TFBatchMonochromaticPSF,
 )
